Tables and lists render with animations
