+++ richtext-ui/src/commonMain/kotlin/com/halilibo/richtext/ui/Table.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clipToBounds
 import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.geometry.Offset
@@ -22,6 +23,8 @@ import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.sp
 import com.halilibo.richtext.ui.ColumnArrangement.Adaptive
 import com.halilibo.richtext.ui.ColumnArrangement.Uniform
+import com.halilibo.richtext.ui.string.MarkdownAnimationState
+import com.halilibo.richtext.ui.string.RichTextRenderOptions
 import kotlin.math.max
 import kotlin.math.roundToInt
 
@@ -112,10 +115,13 @@ private class RowBuilder : RichTextTableCellScope {
 @Composable
 public fun RichTextScope.Table(
   modifier: Modifier = Modifier,
+  markdownAnimationState: MarkdownAnimationState = remember { MarkdownAnimationState() },
+  richTextRenderOptions: RichTextRenderOptions = RichTextRenderOptions(),
   headerRow: (RichTextTableCellScope.() -> Unit)? = null,
   bodyRows: RichTextTableRowScope.() -> Unit
 ) {
   val tableStyle = currentRichTextStyle.resolveDefaults().tableStyle!!
+  val alpha = rememberMarkdownFade(richTextRenderOptions, markdownAnimationState)
   val contentColor = currentContentColor
   val header = remember(headerRow) {
     headerRow?.let { RowBuilder().apply(headerRow).row }
@@ -182,10 +188,11 @@ public fun RichTextScope.Table(
     }
   }
 
+  val baseModifier = modifier.alpha(alpha.value)
   val tableModifier = if (columnArrangement is Adaptive) {
-    modifier.horizontalScroll(rememberScrollState())
+    baseModifier.horizontalScroll(rememberScrollState())
   } else {
-    modifier
+    baseModifier
   }
 
   val borderColor = tableStyle.borderColor!!.takeOrElse { contentColor }
